add bootstrap

+++ src/main/java/org/nhnnext/util/EncryptUtil.java
@@ -27,7 +27,7 @@ public class EncryptUtil {
 
 		bip = md5.digest(tst.getBytes());
 		for (int i = 0; i < bip.length; i++) {
-			eip = "" + Integer.toHexString((int) bip[i] & 0x000000ff);
+			eip = "" + Integer.toHexString(bip[i] & 0x000000ff);
 			if (eip.length() < 2)
 				eip = "0" + eip;
 			temp = temp + eip;
